Rename conversation components and reorganize them
